multi_format_document_analysis/requirements.txt
streamlit
pdfplumber
python-docx
pandas
numpy
matplotlib
scikit-learn
nltk
transformers
torch
wordcloud
